add check for conflicting user supplied autoload file
PHPStan will only load the last autoload-file given on the command-line.
If a user adds "--autoload-file" to the linter args for this plugin
while "use_composer_autoload" is true, and an autoload.php file can be
found, PHPStan will ignore that file silently, which might not be what
a user expects.